move column index to its own module
