+++ parquet/src/file/page_index/index_reader.rs
@@ -18,15 +18,17 @@
 //! Support for reading [`ColumnIndexMetaData`] and [`OffsetIndexMetaData`] from parquet metadata.
 
 use crate::basic::{BoundaryOrder, Type};
-use crate::data_type::private::ParquetValueType;
 use crate::data_type::Int96;
 use crate::errors::{ParquetError, Result};
 use crate::file::metadata::ColumnChunkMetaData;
+use crate::file::page_index::column_index::{
+    ByteArrayColumnIndex, ColumnIndexMetaData, PrimitiveColumnIndex,
+};
 use crate::file::page_index::offset_index::OffsetIndexMetaData;
 use crate::file::reader::ChunkReader;
 use crate::parquet_thrift::{FieldType, ThriftCompactInputProtocol};
 use crate::thrift_struct;
-use std::ops::{Deref, Range};
+use std::ops::Range;
 
 /// Computes the covering range of two optional ranges
 ///
@@ -134,8 +136,9 @@ pub(crate) fn decode_offset_index(data: &[u8]) -> Result<OffsetIndexMetaData, Pa
     OffsetIndexMetaData::try_from(&mut prot)
 }
 
+// private struct only used for decoding then discarded
 thrift_struct!(
-pub(crate) struct ThriftColumnIndex<'a> {
+pub(super) struct ThriftColumnIndex<'a> {
   1: required list<bool> null_pages
   2: required list<'a><binary> min_values
   3: required list<'a><binary> max_values
@@ -145,492 +148,6 @@ pub(crate) struct ThriftColumnIndex<'a> {
   7: optional list<i64> definition_level_histograms;
 }
 );
-
-// TODO: the following should move to its own module
-
-/// Common bits of the column index
-#[derive(Debug, Clone, PartialEq)]
-pub struct ColumnIndex {
-    pub(crate) null_pages: Vec<bool>,
-    pub(crate) boundary_order: BoundaryOrder,
-    pub(crate) null_counts: Option<Vec<i64>>,
-    pub(crate) repetition_level_histograms: Option<Vec<i64>>,
-    pub(crate) definition_level_histograms: Option<Vec<i64>>,
-}
-
-impl ColumnIndex {
-    /// Returns the number of pages
-    pub fn num_pages(&self) -> u64 {
-        self.null_pages.len() as u64
-    }
-
-    /// Returns the number of null values in the page indexed by `idx`
-    ///
-    /// Returns `None` if no null counts have been set in the index
-    pub fn null_count(&self, idx: usize) -> Option<i64> {
-        self.null_counts.as_ref().map(|nc| nc[idx])
-    }
-
-    /// Returns the repetition level histogram for the page indexed by `idx`
-    pub fn repetition_level_histogram(&self, idx: usize) -> Option<&[i64]> {
-        if let Some(rep_hists) = self.repetition_level_histograms.as_ref() {
-            let num_lvls = rep_hists.len() / self.num_pages() as usize;
-            let start = num_lvls * idx;
-            Some(&rep_hists[start..start + num_lvls])
-        } else {
-            None
-        }
-    }
-
-    /// Returns the definition level histogram for the page indexed by `idx`
-    pub fn definition_level_histogram(&self, idx: usize) -> Option<&[i64]> {
-        if let Some(def_hists) = self.definition_level_histograms.as_ref() {
-            let num_lvls = def_hists.len() / self.num_pages() as usize;
-            let start = num_lvls * idx;
-            Some(&def_hists[start..start + num_lvls])
-        } else {
-            None
-        }
-    }
-
-    /// Returns whether the page indexed by `idx` consists of all null values
-    pub fn is_null_page(&self, idx: usize) -> bool {
-        self.null_pages[idx]
-    }
-}
-
-/// Column index for primitive types
-#[derive(Debug, Clone, PartialEq)]
-pub struct PrimitiveColumnIndex<T> {
-    pub(crate) column_index: ColumnIndex,
-    pub(crate) min_values: Vec<T>,
-    pub(crate) max_values: Vec<T>,
-}
-
-impl<T: ParquetValueType> PrimitiveColumnIndex<T> {
-    fn try_new(index: ThriftColumnIndex) -> Result<Self> {
-        let len = index.null_pages.len();
-
-        let mut min_values = Vec::with_capacity(len);
-        let mut max_values = Vec::with_capacity(len);
-
-        for (i, is_null) in index.null_pages.iter().enumerate().take(len) {
-            if !is_null {
-                let min = index.min_values[i];
-                min_values.push(T::try_from_le_slice(min)?);
-
-                let max = index.max_values[i];
-                max_values.push(T::try_from_le_slice(max)?);
-            } else {
-                // need placeholders
-                min_values.push(Default::default());
-                max_values.push(Default::default());
-            }
-        }
-
-        Ok(Self {
-            column_index: ColumnIndex {
-                null_pages: index.null_pages,
-                boundary_order: index.boundary_order,
-                null_counts: index.null_counts,
-                repetition_level_histograms: index.repetition_level_histograms,
-                definition_level_histograms: index.definition_level_histograms,
-            },
-            min_values,
-            max_values,
-        })
-    }
-
-    pub(crate) fn to_thrift(&self) -> crate::format::ColumnIndex {
-        let min_values = self
-            .min_values
-            .iter()
-            .map(|x| x.as_bytes().to_vec())
-            .collect::<Vec<_>>();
-
-        let max_values = self
-            .max_values
-            .iter()
-            .map(|x| x.as_bytes().to_vec())
-            .collect::<Vec<_>>();
-
-        let null_counts = self.null_counts.clone();
-        let repetition_level_histograms = self.repetition_level_histograms.clone();
-        let definition_level_histograms = self.definition_level_histograms.clone();
-        let null_pages = self.null_pages.clone();
-
-        crate::format::ColumnIndex::new(
-            null_pages,
-            min_values,
-            max_values,
-            self.boundary_order.into(),
-            null_counts,
-            repetition_level_histograms,
-            definition_level_histograms,
-        )
-    }
-}
-
-impl<T> PrimitiveColumnIndex<T> {
-    /// Returns an array containing the min values for each page.
-    ///
-    /// Values in the returned slice are only valid if [`ColumnIndex::is_null_page()`]
-    /// is `false` for the same index.
-    pub fn min_values(&self) -> &[T] {
-        &self.min_values
-    }
-
-    /// Returns an array containing the max values for each page.
-    ///
-    /// Values in the returned slice are only valid if [`ColumnIndex::is_null_page()`]
-    /// is `false` for the same index.
-    pub fn max_values(&self) -> &[T] {
-        &self.max_values
-    }
-
-    /// Returns an iterator over the min values.
-    ///
-    /// Values may be `None` when [`ColumnIndex::is_null_page()`] is `true`.
-    pub fn min_values_iter(&self) -> impl Iterator<Item = Option<&T>> {
-        self.min_values.iter().enumerate().map(|(i, min)| {
-            if self.is_null_page(i) {
-                None
-            } else {
-                Some(min)
-            }
-        })
-    }
-
-    /// Returns an iterator over the max values.
-    ///
-    /// Values may be `None` when [`ColumnIndex::is_null_page()`] is `true`.
-    pub fn max_values_iter(&self) -> impl Iterator<Item = Option<&T>> {
-        self.max_values.iter().enumerate().map(|(i, min)| {
-            if self.is_null_page(i) {
-                None
-            } else {
-                Some(min)
-            }
-        })
-    }
-
-    /// Returns the min value for the page indexed by `idx`
-    ///
-    /// It is `None` when all values are null
-    pub fn min_value(&self, idx: usize) -> Option<&T> {
-        if self.null_pages[idx] {
-            None
-        } else {
-            Some(&self.min_values[idx])
-        }
-    }
-
-    /// Returns the max value for the page indexed by `idx`
-    ///
-    /// It is `None` when all values are null
-    pub fn max_value(&self, idx: usize) -> Option<&T> {
-        if self.null_pages[idx] {
-            None
-        } else {
-            Some(&self.max_values[idx])
-        }
-    }
-}
-
-impl<T> Deref for PrimitiveColumnIndex<T> {
-    type Target = ColumnIndex;
-
-    fn deref(&self) -> &Self::Target {
-        &self.column_index
-    }
-}
-
-/// Column index for byte arrays (fixed length and variable)
-#[derive(Debug, Clone, PartialEq)]
-pub struct ByteArrayColumnIndex {
-    pub(crate) column_index: ColumnIndex,
-    // raw bytes for min and max values
-    pub(crate) min_bytes: Vec<u8>,
-    pub(crate) min_offsets: Vec<usize>,
-    pub(crate) max_bytes: Vec<u8>,
-    pub(crate) max_offsets: Vec<usize>,
-}
-
-impl ByteArrayColumnIndex {
-    fn try_new(index: ThriftColumnIndex) -> Result<Self> {
-        let len = index.null_pages.len();
-
-        let min_len = index.min_values.iter().map(|&v| v.len()).sum();
-        let max_len = index.max_values.iter().map(|&v| v.len()).sum();
-        let mut min_bytes = vec![0u8; min_len];
-        let mut max_bytes = vec![0u8; max_len];
-
-        let mut min_offsets = vec![0usize; len + 1];
-        let mut max_offsets = vec![0usize; len + 1];
-
-        let mut min_pos = 0;
-        let mut max_pos = 0;
-
-        for (i, is_null) in index.null_pages.iter().enumerate().take(len) {
-            if !is_null {
-                let min = index.min_values[i];
-                let dst = &mut min_bytes[min_pos..min_pos + min.len()];
-                dst.copy_from_slice(min);
-                min_offsets[i] = min_pos;
-                min_pos += min.len();
-
-                let max = index.max_values[i];
-                let dst = &mut max_bytes[max_pos..max_pos + max.len()];
-                dst.copy_from_slice(max);
-                max_offsets[i] = max_pos;
-                max_pos += max.len();
-            } else {
-                min_offsets[i] = min_pos;
-                max_offsets[i] = max_pos;
-            }
-        }
-
-        min_offsets[len] = min_pos;
-        max_offsets[len] = max_pos;
-
-        Ok(Self {
-            column_index: ColumnIndex {
-                null_pages: index.null_pages,
-                boundary_order: index.boundary_order,
-                null_counts: index.null_counts,
-                repetition_level_histograms: index.repetition_level_histograms,
-                definition_level_histograms: index.definition_level_histograms,
-            },
-
-            min_bytes,
-            min_offsets,
-            max_bytes,
-            max_offsets,
-        })
-    }
-
-    /// Returns the min value for the page indexed by `idx`
-    ///
-    /// It is `None` when all values are null
-    pub fn min_value(&self, idx: usize) -> Option<&[u8]> {
-        if self.null_pages[idx] {
-            None
-        } else {
-            let start = self.min_offsets[idx];
-            let end = self.min_offsets[idx + 1];
-            Some(&self.min_bytes[start..end])
-        }
-    }
-
-    /// Returns the max value for the page indexed by `idx`
-    ///
-    /// It is `None` when all values are null
-    pub fn max_value(&self, idx: usize) -> Option<&[u8]> {
-        if self.null_pages[idx] {
-            None
-        } else {
-            let start = self.max_offsets[idx];
-            let end = self.max_offsets[idx + 1];
-            Some(&self.max_bytes[start..end])
-        }
-    }
-
-    /// Returns an iterator over the min values.
-    ///
-    /// Values may be `None` when [`ColumnIndex::is_null_page()`] is `true`.
-    pub fn min_values_iter(&self) -> impl Iterator<Item = Option<&[u8]>> {
-        (0..self.num_pages() as usize).map(|i| {
-            if self.is_null_page(i) {
-                None
-            } else {
-                self.min_value(i)
-            }
-        })
-    }
-
-    /// Returns an iterator over the max values.
-    ///
-    /// Values may be `None` when [`ColumnIndex::is_null_page()`] is `true`.
-    pub fn max_values_iter(&self) -> impl Iterator<Item = Option<&[u8]>> {
-        (0..self.num_pages() as usize).map(|i| {
-            if self.is_null_page(i) {
-                None
-            } else {
-                self.max_value(i)
-            }
-        })
-    }
-
-    pub(crate) fn to_thrift(&self) -> crate::format::ColumnIndex {
-        let mut min_values = Vec::with_capacity(self.num_pages() as usize);
-        for i in 0..self.num_pages() as usize {
-            min_values.push(self.min_value(i).unwrap_or(&[]).to_owned());
-        }
-
-        let mut max_values = Vec::with_capacity(self.num_pages() as usize);
-        for i in 0..self.num_pages() as usize {
-            max_values.push(self.max_value(i).unwrap_or(&[]).to_owned());
-        }
-
-        let null_counts = self.null_counts.clone();
-        let repetition_level_histograms = self.repetition_level_histograms.clone();
-        let definition_level_histograms = self.definition_level_histograms.clone();
-        let null_pages = self.null_pages.clone();
-
-        crate::format::ColumnIndex::new(
-            null_pages,
-            min_values,
-            max_values,
-            self.boundary_order.into(),
-            null_counts,
-            repetition_level_histograms,
-            definition_level_histograms,
-        )
-    }
-}
-
-impl Deref for ByteArrayColumnIndex {
-    type Target = ColumnIndex;
-
-    fn deref(&self) -> &Self::Target {
-        &self.column_index
-    }
-}
-
-// Macro to generate getter functions for ColumnIndexMetaData.
-macro_rules! colidx_enum_func {
-    ($self:ident, $func:ident, $arg:ident) => {{
-        match *$self {
-            Self::BOOLEAN(ref typed) => typed.$func($arg),
-            Self::INT32(ref typed) => typed.$func($arg),
-            Self::INT64(ref typed) => typed.$func($arg),
-            Self::INT96(ref typed) => typed.$func($arg),
-            Self::FLOAT(ref typed) => typed.$func($arg),
-            Self::DOUBLE(ref typed) => typed.$func($arg),
-            Self::BYTE_ARRAY(ref typed) => typed.$func($arg),
-            Self::FIXED_LEN_BYTE_ARRAY(ref typed) => typed.$func($arg),
-            _ => panic!(concat!(
-                "Cannot call ",
-                stringify!($func),
-                " on ColumnIndexMetaData::NONE"
-            )),
-        }
-    }};
-    ($self:ident, $func:ident) => {{
-        match *$self {
-            Self::BOOLEAN(ref typed) => typed.$func(),
-            Self::INT32(ref typed) => typed.$func(),
-            Self::INT64(ref typed) => typed.$func(),
-            Self::INT96(ref typed) => typed.$func(),
-            Self::FLOAT(ref typed) => typed.$func(),
-            Self::DOUBLE(ref typed) => typed.$func(),
-            Self::BYTE_ARRAY(ref typed) => typed.$func(),
-            Self::FIXED_LEN_BYTE_ARRAY(ref typed) => typed.$func(),
-            _ => panic!(concat!(
-                "Cannot call ",
-                stringify!($func),
-                " on ColumnIndexMetaData::NONE"
-            )),
-        }
-    }};
-}
-
-/// index
-#[derive(Debug, Clone, PartialEq)]
-#[allow(non_camel_case_types)]
-pub enum ColumnIndexMetaData {
-    /// Sometimes reading page index from parquet file
-    /// will only return pageLocations without min_max index,
-    /// `NONE` represents this lack of index information
-    NONE,
-    /// Boolean type index
-    BOOLEAN(PrimitiveColumnIndex<bool>),
-    /// 32-bit integer type index
-    INT32(PrimitiveColumnIndex<i32>),
-    /// 64-bit integer type index
-    INT64(PrimitiveColumnIndex<i64>),
-    /// 96-bit integer type (timestamp) index
-    INT96(PrimitiveColumnIndex<Int96>),
-    /// 32-bit floating point type index
-    FLOAT(PrimitiveColumnIndex<f32>),
-    /// 64-bit floating point type index
-    DOUBLE(PrimitiveColumnIndex<f64>),
-    /// Byte array type index
-    BYTE_ARRAY(ByteArrayColumnIndex),
-    /// Fixed length byte array type index
-    FIXED_LEN_BYTE_ARRAY(ByteArrayColumnIndex),
-}
-
-impl ColumnIndexMetaData {
-    /// Return min/max elements inside ColumnIndex are ordered or not.
-    pub fn is_sorted(&self) -> bool {
-        // 0:UNORDERED, 1:ASCENDING ,2:DESCENDING,
-        if let Some(order) = self.get_boundary_order() {
-            order != BoundaryOrder::UNORDERED
-        } else {
-            false
-        }
-    }
-
-    /// Get boundary_order of this page index.
-    pub fn get_boundary_order(&self) -> Option<BoundaryOrder> {
-        match self {
-            Self::NONE => None,
-            Self::BOOLEAN(index) => Some(index.boundary_order),
-            Self::INT32(index) => Some(index.boundary_order),
-            Self::INT64(index) => Some(index.boundary_order),
-            Self::INT96(index) => Some(index.boundary_order),
-            Self::FLOAT(index) => Some(index.boundary_order),
-            Self::DOUBLE(index) => Some(index.boundary_order),
-            Self::BYTE_ARRAY(index) => Some(index.boundary_order),
-            Self::FIXED_LEN_BYTE_ARRAY(index) => Some(index.boundary_order),
-        }
-    }
-
-    /// Returns array of null counts, one per page.
-    ///
-    /// Returns `None` if now null counts have been set in the index
-    pub fn null_counts(&self) -> Option<&Vec<i64>> {
-        match self {
-            Self::NONE => None,
-            Self::BOOLEAN(index) => index.null_counts.as_ref(),
-            Self::INT32(index) => index.null_counts.as_ref(),
-            Self::INT64(index) => index.null_counts.as_ref(),
-            Self::INT96(index) => index.null_counts.as_ref(),
-            Self::FLOAT(index) => index.null_counts.as_ref(),
-            Self::DOUBLE(index) => index.null_counts.as_ref(),
-            Self::BYTE_ARRAY(index) => index.null_counts.as_ref(),
-            Self::FIXED_LEN_BYTE_ARRAY(index) => index.null_counts.as_ref(),
-        }
-    }
-
-    /// Returns the number of pages
-    pub fn num_pages(&self) -> u64 {
-        colidx_enum_func!(self, num_pages)
-    }
-
-    /// Returns the number of null values in the page indexed by `idx`
-    ///
-    /// Returns `None` if no null counts have been set in the index
-    pub fn null_count(&self, idx: usize) -> Option<i64> {
-        colidx_enum_func!(self, null_count, idx)
-    }
-
-    /// Returns the repetition level histogram for the page indexed by `idx`
-    pub fn repetition_level_histogram(&self, idx: usize) -> Option<&[i64]> {
-        colidx_enum_func!(self, repetition_level_histogram, idx)
-    }
-
-    /// Returns the definition level histogram for the page indexed by `idx`
-    pub fn definition_level_histogram(&self, idx: usize) -> Option<&[i64]> {
-        colidx_enum_func!(self, definition_level_histogram, idx)
-    }
-
-    /// Returns whether the page indexed by `idx` consists of all null values
-    pub fn is_null_page(&self, idx: usize) -> bool {
-        colidx_enum_func!(self, is_null_page, idx)
-    }
-}
 
 pub(crate) fn decode_column_index(
     data: &[u8],
