+++ parquet/src/file/page_index/mod.rs
@@ -19,6 +19,7 @@
 //!
 //! [Column Index]: https://github.com/apache/parquet-format/blob/master/PageIndex.md
 
+pub mod column_index;
 pub mod index;
 pub mod index_reader;
 pub mod offset_index;
